+++ KeywordsApp/Models/Keyword/KeywordResultParserModel.cs
@@ -15,7 +15,7 @@ namespace KeywordsApp.Models.Keyword
         public int AdWordsCount { get; set; }
         public int LinkCount { get; set; }
 
-        public int TotalThouthandResultsCount { get; set; }
+        public double TotalThouthandResultsCount { get; set; }
         public string HtmlCode { get; set; }
 
         // Miliseconds
